+++ packages/core/src/utils/filesearch/crawler.test.ts
@@ -9,7 +9,10 @@ import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 import * as cache from './crawlCache.js';
 import { crawl } from './crawler.js';
-import { createTmpDir, cleanupTmpDir } from '@qwen-code/qwen-code-test-utils';
+import {
+  createTmpDir,
+  cleanupTmpDir,
+} from '../../test-utils/file-system-test-helpers.js';
 import type { Ignore } from './ignore.js';
 import { loadIgnoreRules } from './ignore.js';
