+++ packages/core/src/utils/filesearch/fileSearch.test.ts
@@ -6,7 +6,10 @@
 
 import { describe, it, expect, afterEach, vi } from 'vitest';
 import { FileSearchFactory, AbortError, filter } from './fileSearch.js';
-import { createTmpDir, cleanupTmpDir } from '@qwen-code/qwen-code-test-utils';
+import {
+  createTmpDir,
+  cleanupTmpDir,
+} from '../../test-utils/file-system-test-helpers.js';
 
 describe('FileSearch', () => {
   let tmpDir: string;
