--- packages/test-utils/index.ts
@@ -1,7 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './src/file-system-test-helpers.js';
