--- packages/test-utils/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@qwen-code/qwen-code-test-utils",
-  "version": "0.14.3",
-  "private": true,
-  "main": "src/index.ts",
-  "license": "Apache-2.0",
-  "type": "module",
-  "scripts": {
-    "build": "node ../../scripts/build_package.js",
-    "typecheck": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "typescript": "^5.3.3"
-  },
-  "engines": {
-    "node": ">=20"
-  }
-}
