--- packages/test-utils/vitest.config.ts
@@ -1,23 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  test: {
-    reporters: ['default', 'junit'],
-    silent: true,
-    outputFile: {
-      junit: 'junit.xml',
-    },
-    poolOptions: {
-      threads: {
-        minThreads: 8,
-        maxThreads: 16,
-      },
-    },
-  },
-});
